fit: float64 NLL reduction in step1/step2 (adam/soap/lbfgs paths)
Upcast the per-event log-density / NLL to float64 before the Σw-weighted
reduction in step1 (flow) and step2 (θ fit), so the within-batch sum AND its
backward run in float64 while the model stays at --precision — same treatment
the trust-region driver's _per_event_term already uses, now applied to every
optimiser that goes through step_fn (adam, soap, lbfgs). A ~65k-event float32
sum carries ~√N·ε ≈ 3e-5 relative cancellation; the float64 reduction removes
it, and the backward accumulates per-event grad contributions in float64, cast
to the fp32 param leaf only at the end (mixed-dtype autograd).

Cross-batch: the adam/soap minibatch loop steps PER batch (no cross-batch
gradient accumulation — only the reported tr_sum sums across batches, already a
Python float64). The L-BFGS closure DOES accumulate gradient across batches, but
into the fp32 .grad leaf (a float64 .grad would need a float64 model); its scalar
objective sum is already float64. So the substantive change for these paths is
the within-batch float64 reduction, which they now all get.

Verified: adam, soap, and lbfgs all train cleanly with the float64-loss reduction
(mixed float64-loss / fp32-param backward; θ moving, NLL descending, EXIT=0).

+++ scripts/corrections/muon_calibration/train_jpsi_mass_fit.py
@@ -1402,11 +1402,15 @@ def train_stage1(args, model, train_loader, val_loader, stats) -> float:
     def step1(model, batch):
         idx = (~batch["is_data_mask"]).nonzero(as_tuple=True)[0]
         if idx.numel() == 0:
-            return torch.zeros((), device=batch["mll"].device), 0.0
+            return torch.zeros((), dtype=torch.float64,
+                               device=batch["mll"].device), 0.0
         logp = model.log_p_nominal(batch["mll"][idx], batch["cond_std"][idx])
-        w = batch["w"][idx]
-        sw = float(w.sum().clamp_min(1e-30))
-        return -(w * logp).sum() / sw, sw
+        w = batch["w"][idx].double()                 # float64 reduction: the model
+        sw = float(w.sum().clamp_min(1e-30))         # runs at --precision, but the
+        return -(w * logp.double()).sum() / sw, sw   # Σw·NLL sum + its backward are
+        # float64 (a ~65k-event float32 sum carries ~√N·ε cancellation; backward
+        # accumulates per-event grad contributions in float64, cast to the fp32
+        # leaf only at the end). Benefits adam/soap/lbfgs alike (all call step_fn).
 
     return _run_epochs(args, model, optim, train_loader, val_loader, stats,
                        step_fn=step1, ckpt_prefix="flow", stage_name="flow",
@@ -1483,9 +1487,12 @@ def train_stage2(args, model, train_loader, val_loader, stats,
             batch["mll"], batch["pt_pm"], batch["eta_pm"], batch["phi_pm"],
             batch["q_pm"], batch["b_pm"], batch["cond_std"], data_mask,
             n_iter=args.continuity_n_iter)
-        w = batch["w"] * data_mask.to(batch["w"].dtype)
+        # float64 Σw·NLL reduction + backward (model stays at --precision); same
+        # rationale as step1 / the trust driver's _per_event_term, applied to all
+        # optimisers that go through step_fn (adam/soap/lbfgs).
+        w = (batch["w"] * data_mask.to(batch["w"].dtype)).double()
         sw = float(w.sum().clamp_min(1e-30))
-        return (w * per).sum() / sw, sw
+        return (w * per.double()).sum() / sw, sw
 
     fit_opt = getattr(args, "fit_optimizer", "adam")
     max_epochs = args.fit_epochs or args.epochs
